requirements.txt
aiocache==0.12.1
aiofiles==23.1.0
aiogoogletrans==3.3.3
aiohttp==3.8.4
aiosqlite==0.19.0
akinator==1.0.3
async_cleverbot==0.2.2
async_cse==0.3.0
asyncdagpi==4.1.5
asyncpg==0.27.0
asyncpraw==7.7.0
beautifulsoup4==4.12.2
chess==1.9.4
colorama==0.4.6
discord.py==2.2.3
easy_pil==0.2.4
emojis==0.7.0
feedparser==6.0.10
gradio_tools==0.0.9
gtts==2.3.2
httpx==0.24.0
humanfriendly==10.0
humanize==4.6.0
joblib==1.2.0
langchain==0.0.171
langdetect==1.0.9
loguru==0.7.0
matplotlib==3.7.1
mcstatus==10.0.3
nextcord==2.4.2
nltk==3.8.1
numpy==1.24.3
openai==0.27.6
owotext==1.1.1
Pillow==9.5.0
playsound==1.3.0
protobuf==4.23.0
psutil==5.9.5
pycoingecko==3.1.0
pydub==0.25.1
pyparsing==3.0.9
python-dotenv==1.0.0
python_dateutil==2.8.2
pyttsx3==2.90
pytube==15.0.0
pytz==2023.3
regex==2023.5.5
Requests==2.30.0
rich==13.3.5
scikit_learn==1.2.2
sourceserver==0.9.2
tqdm==4.65.0
tweepy==4.14.0
typing_extensions==4.5.0
validators==0.20.0
wolframalpha==5.0.0
youtube_dl==2021.12.17
